requirements.txt
altair==5.0.1
asttokens==2.4.1
attrs==23.1.0
audioread==3.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
ffmpeg==1.4
ffmpeg-python==0.2.0
streamlit
Flask==3.0.3
fonttools==4.51.0
future==1.0.0
gitdb==4.0.10
GitPython==3.1.31
gTTS==2.5.1
idna==3.4
importlib-metadata==6.7.0
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.0.8
numba==0.59.1
numpy==1.25.0
opencv-python-headless==4.9.0.80
packaging==23.1
pandas==2.0.3
parso==0.8.4
Pillow==9.5.0
platformdirs==4.2.1
pooch==1.8.1
prompt-toolkit==3.0.43
protobuf==4.23.4
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.22
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.29.1
requests==2.31.0
rich==13.4.2
rpds-py==0.8.8
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
streamlit==1.24.1
streamlit-option-menu==0.3.12
tenacity==8.2.2
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
traitlets==5.14.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.15.0
